+++ lib/src/config.rs
@@ -39,6 +39,7 @@ pub use crate::config_resolver::ConfigMigrateError;
 pub use crate::config_resolver::ConfigMigrateLayerError;
 pub use crate::config_resolver::ConfigMigrationRule;
 pub use crate::config_resolver::ConfigResolutionContext;
+pub use crate::config_resolver::RemoteInfo;
 pub use crate::config_resolver::migrate;
 pub use crate::config_resolver::resolve;
 use crate::file_util::IoResultExt as _;